demo-1/backend/requirements.txt
flask
pycryptodome
psutil
pandas
matplotlib
tqdm
